backend/requirements-dev.txt
pytest
pytest-asyncio
httpx
coverage
psycopg2-binary
